Moving report folder and timeout definition to config
see #117
