+++ galen-junit/src/main/java/net/mindengine/galen/junit/JUnitStepListener.java
@@ -18,6 +18,8 @@ package net.mindengine.galen.junit;
 import java.util.List;
 
 import net.mindengine.galen.api.GalenReportsContainer;
+import net.mindengine.galen.config.GalenConfig;
+import net.mindengine.galen.config.GalenProperty;
 import net.mindengine.galen.reports.GalenTestInfo;
 import net.mindengine.galen.reports.HtmlReportBuilder;
 import net.mindengine.galen.reports.model.FileTempStorage;
@@ -43,8 +45,7 @@ public class JUnitStepListener extends RunListener {
         LOG.info("Generating Galen Html reports");
         List<GalenTestInfo> tests = GalenReportsContainer.get().getAllTests();
         try {
-            // TODO move to variable
-            new HtmlReportBuilder().build(tests, "target/galen-html-reports");
+            new HtmlReportBuilder().build(tests, GalenConfig.getConfig().readProperty(GalenProperty.TEST_JAVA_REPORT_OUTPUTFOLDER));
             cleanData(tests);
         } catch (Exception e) {
             throw new RuntimeException(e);
